:art: https://github.com/siyuan-note/siyuan/issues/10749
