+++ app/src/window/init.ts
@@ -2,7 +2,7 @@ import {Constants} from "../constants";
 import {ipcRenderer, webFrame} from "electron";
 import {fetchPost} from "../util/fetch";
 import {adjustLayout, getInstanceById, JSONToCenter} from "../layout/util";
-import {resizeTabs} from "../layout/tabUtil";
+import {resizeTabs, setTabPosition} from "../layout/tabUtil";
 import {initStatus} from "../layout/status";
 import {appearance} from "../config/appearance";
 import {initAssets, setInlineStyle} from "../util/assets";
@@ -12,12 +12,12 @@ import {initWindow} from "../boot/onGetConfig";
 import {App} from "../index";
 import {afterLoadPlugin} from "../plugin/loader";
 import {Tab} from "../layout/Tab";
-import {initWindowEvent} from "../boot/globalEvent/event";
-import {getAllEditor} from "../layout/getAll";
 /// #if !BROWSER
 import {initNativeDialogOverride} from "../protyle/util/compatibility";
-import {setTabPosition} from "./setHeader";
 /// #endif
+import {initWindowEvent} from "../boot/globalEvent/event";
+import {getAllEditor} from "../layout/getAll";
+
 
 export const init = (app: App) => {
     webFrame.setZoomFactor(window.siyuan.storage[Constants.LOCAL_ZOOM]);
